Update rewards.py
Updated the txid saving to the database on each transaction.
Added logic to check for existing txid's and then continue an existing distribution of rewards.

Added additional logic to check the database for existing asset_names, addresses and quantities when adding new asset names to avoid duplication when fetching asset information.

Added progress bar to show transfer progress when distributing rewards. Progress resets to 0 when restarting a distribution if it failed for some reason. Next run ignores database entries that have been already rewarded and have a txid.