Add T2_ipfs_pinner_0_0_3_beta tool
Added an early beta tool I've been working on.

The script takes some user input and setup and in intended to help most computer minded people setup the Ravencoin and IPFS node local communications over RPC. 

When running the script, it will ask the user questions and setup a config file specific to that machine and user.

This script should work on UNIX and windows systems. I specifically avoided using imports that are not cross compatible.

The intent is to assist users to pin IPFS content directly from their Ravencoin node to their local IPFS node. It does this by creating a sqlite database of asset information scraped from the local Ravencoin node and then using that database information to download the IPFS hashes to the local IPFS node. The script checks for the IPFS file size and if within configuration settings adds and pins the IPFS content to the local IPFS node.

Install the required depends using the below. On windows systems ensure python is installed.

Tested using Python version 3.10. 

Ensure the below dependancies are available.

pip install bitcoinrpc
pip install sqlite3
in future pyzmq will be need but not at the moment as ZMQ code is not functional and tested as yet.